Sema: warn when POSITION[0] is used intead of SV_Position (#6401)
Before SM4, POSITION was equivalent to today's SV_Position. However, as
HLSL moved to system semantics (SV_), the support for those old
semantics has been deprecated.
Some like VFACE are now completely forbidden, but others like POSITION
or COLOR have been handled differently by FXC and DXC.

Today, usage of POSITION in place of SV_Position is probably a mistake
as POSITION is now considered to be a user-defined semantic.

This commit adds a warning when POSITION is used as an output semantic
for a vertex shader, as most cases should be the result of a mistake. If
this is expected, the warning can be disabled by using
`-Wno-dx9-deprecation`.

Fixes #3742

---------

Signed-off-by: Nathan Gauër <brioche@google.com>
(cherry picked from commit eee0416c77ea0384cd0d0d69aa7fdca979471d62)